extracted media-selection-item into seperate component, use media-selection-item in single-media-selection
