+++ src/Sulu/Bundle/MediaBundle/Resources/js/components/MediaSelectionItem/index.js
@@ -1,0 +1,4 @@
+// @flow
+import MediaSelectionItem from './MediaSelectionItem';
+
+export default MediaSelectionItem;
