+++ src/Sulu/Bundle/MediaBundle/Resources/js/components/MediaSelectionItem/mediaSelectionItem.scss
